remove Block Manager modal from options
