+++ packages/edit-post/src/plugins/index.js
@@ -11,7 +11,6 @@ import { addQueryArgs } from '@wordpress/url';
  * Internal dependencies
  */
 import CopyContentMenuItem from './copy-content-menu-item';
-import ManageBlocksMenuItem from './manage-blocks-menu-item';
 import KeyboardShortcutsHelpMenuItem from './keyboard-shortcuts-help-menu-item';
 import ToolsMoreMenuGroup from '../components/header/tools-more-menu-group';
 import WelcomeGuideMenuItem from './welcome-guide-menu-item';
@@ -23,7 +22,6 @@ registerPlugin( 'edit-post', {
 				<ToolsMoreMenuGroup>
 					{ ( { onClose } ) => (
 						<>
-							<ManageBlocksMenuItem onSelect={ onClose } />
 							<MenuItem
 								role="menuitem"
 								href={ addQueryArgs( 'edit.php', {
